delete bin/
`bin/setup` and `bin/update` are examples scripts that teams can have
to help developers get their development environments up and running.

The rest are binstubs used for preloading the app.

Now to start up the server, we'll have to use Rack directly with the
`config.ru` file using the `rackup` command. From now on, the server
will default of port 9292 instead of 3000.