"404 as a service" -- Routes? Who needs routes?
We don't need to render anything in order to test if the server is
running. As long as there is a connection to the server and there is
any response, we can confirm the app is running.

Instead of using the browser, `curl` to the rescue!
```
$ curl --head localhost:9292
HTTP/1.1 404 Not Found
Content-Type: text/html; charset=utf-8
X-Request-Id: f9b0ac3c-bed4-4c08-aa92-759f839bafe3
X-Runtime: 0.001348
Content-Length: 0
```